ExproniconLite

ExproniconLite.no_default

ExproniconLite.JLExpr

ExproniconLite.JLField

ExproniconLite.JLField

ExproniconLite.JLFor

ExproniconLite.JLFor

ExproniconLite.JLFor

ExproniconLite.JLFunction

ExproniconLite.JLFunction

ExproniconLite.JLIfElse

ExproniconLite.JLIfElse

ExproniconLite.JLIfElse

ExproniconLite.JLKwField

ExproniconLite.JLKwField

ExproniconLite.JLKwStruct

ExproniconLite.JLKwStruct

ExproniconLite.JLKwStruct

ExproniconLite.JLStruct

ExproniconLite.JLStruct

ExproniconLite.JLStruct

ExproniconLite.NoDefault

ExproniconLite.codegen_ast

ExproniconLite.codegen_ast_fields

ExproniconLite.codegen_ast_kwfn

ExproniconLite.codegen_ast_kwfn_infer

ExproniconLite.codegen_ast_kwfn_plain

ExproniconLite.codegen_ast_struct

ExproniconLite.codegen_ast_struct_body

ExproniconLite.codegen_ast_struct_head

ExproniconLite.compare_expr

ExproniconLite.compare_vars

ExproniconLite.eval_interp

ExproniconLite.eval_literal

ExproniconLite.flatten_blocks

ExproniconLite.has_kwfn_constructor

ExproniconLite.has_plain_constructor

ExproniconLite.is_field

ExproniconLite.is_field_default

ExproniconLite.is_for

ExproniconLite.is_function

ExproniconLite.is_ifelse

ExproniconLite.is_kw_function

ExproniconLite.is_literal

ExproniconLite.is_struct

ExproniconLite.is_struct_not_kw_struct

ExproniconLite.name_only

ExproniconLite.prettify

ExproniconLite.rm_annotations

ExproniconLite.rm_lineinfo

ExproniconLite.rm_nothing

ExproniconLite.split_doc

ExproniconLite.split_function

ExproniconLite.split_function_head

ExproniconLite.split_struct

ExproniconLite.split_struct_name

ExproniconLite.struct_name_plain

ExproniconLite.struct_name_without_inferable

ExproniconLite.subtitute

ExproniconLite.support_default

ExproniconLite.xcall

ExproniconLite.xcall

ExproniconLite.xfirst

ExproniconLite.xiterate

ExproniconLite.xiterate

ExproniconLite.xlast

ExproniconLite.xmap

ExproniconLite.xmapreduce

ExproniconLite.xnamedtuple

ExproniconLite.xprint

ExproniconLite.xprintln

ExproniconLite.xpush

ExproniconLite.xtuple

ExproniconLite.@expr

ExproniconLite.@expr

ExproniconLite.@test_expr

ExproniconLite.@test_expr

ExproniconLite.no_default — Constant

const no_default = NoDefault()

Constant instance for NoDefault that describes a field should have no default value.

ExproniconLite.JLExpr — Type

abstract type JLExpr end

Abstract type for Julia syntax type.

ExproniconLite.JLField — Type

JLField <: JLExpr

Type describes a Julia field in a Julia struct.

ExproniconLite.JLField — Method

JLField(;name, type=Any, doc=nothing, line=nothing)

Create a JLField instance.

ExproniconLite.JLFor — Type

JLFor <: JLExpr

Syntax type for Julia for loop.

ExproniconLite.JLFor — Method

JLFor(ex::Expr)

Create a JLFor from given Julia for loop expression.

Example

julia> ex = @expr for i in 1:10, j in 1:j
           M[i, j] += 1
       end
:(for i = 1:10, j = 1:j
      #= REPL[3]:2 =#
      M[i, j] += 1
  end)

julia> jl = JLFor(ex)
for i in 1 : 10,
    j in 1 : j
    #= loop body =#
    begin
        #= REPL[3]:2 =#        
        M[i, j] += 1        
    end
end

julia> jl.vars
2-element Vector{Any}:
 :i
 :j

julia> jl.iterators
2-element Vector{Any}:
 :(1:10)
 :(1:j)

ExproniconLite.JLFor — Method

JLFor(;vars=[], iterators=[], kernel=nothing)

Generate a JLFor object.

Kwargs

vars: loop variables.

iterators: loop iterators.

kernel: loop kernel.

ExproniconLite.JLFunction — Type

JLFunction <: JLExpr

Type describes a Julia function declaration expression.

Example

Construct a function expression

julia> JLFunction(;name=:foo, args=[:(x::T)], body= quote 1+1 end, head=:function, whereparams=[:T])
function foo(x::T) where {T}
    #= REPL[25]:1 =#    
    1 + 1    
end

Decompose a function expression

julia> ex = :(function foo(x::T) where {T}
           #= REPL[25]:1 =#    
           1 + 1    
       end)
:(function foo(x::T) where T
      #= REPL[26]:1 =#
      #= REPL[26]:3 =#
      1 + 1
  end)

julia> jl = JLFunction(ex)
function foo(x::T) where {T}
    #= REPL[26]:1 =#    
    #= REPL[26]:3 =#    
    1 + 1    
end

Generate Expr from JLFunction

julia> codegen_ast(jl)
:(function foo(x::T) where T
      #= REPL[26]:1 =#
      #= REPL[26]:3 =#
      1 + 1
  end)

ExproniconLite.JLFunction — Method

JLFunction(ex::Expr)

Create a JLFunction object from a Julia function Expr.

Example

julia> JLFunction(:(f(x) = 2))
f(x) = begin
    #= REPL[37]:1 =#    
    2    
end

ExproniconLite.JLIfElse — Type

JLIfElse <: JLExpr

JLIfElse describes a Julia if ... elseif ... else ... end expression. It allows one to easily construct such expression by inserting condition and code block via a map.

Example

Construct JLIfElse object

One can construct an ifelse as following

julia> jl = JLIfElse()
nothing

julia> jl.map[:(foo(x))] = :(x = 1 + 1)
:(x = 1 + 1)

julia> jl.map[:(goo(x))] = :(y = 1 + 2)
:(y = 1 + 2)

julia> jl.otherwise = :(error("abc"))
:(error("abc"))

julia> jl
if foo(x)
    x = 1 + 1
elseif goo(x)
    y = 1 + 2
else
    error("abc")
end

Generate the Julia Expr object

to generate the corresponding Expr object, one can call codegen_ast.

julia> codegen_ast(jl)
:(if foo(x)
      x = 1 + 1
  elseif goo(x)
      y = 1 + 2
  else
      error("abc")
  end)

ExproniconLite.JLIfElse — Method

JLIfElse(ex::Expr)

Create a JLIfElse from given Julia ifelse Expr.

Example

julia> ex = :(if foo(x)
             x = 1 + 1
         elseif goo(x)
             y = 1 + 2
         else
             error("abc")
         end)
:(if foo(x)
      #= REPL[41]:2 =#
      x = 1 + 1
  elseif #= REPL[41]:3 =# goo(x)
      #= REPL[41]:4 =#
      y = 1 + 2
  else
      #= REPL[41]:6 =#
      error("abc")
  end)

julia> JLIfElse(ex)
if foo(x)
    begin
        #= REPL[41]:2 =#        
        x = 1 + 1        
    end
elseif begin
    #= REPL[41]:3 =#    
    goo(x)    
end
    begin
        #= REPL[41]:4 =#        
        y = 1 + 2        
    end
else
    begin
        #= REPL[41]:6 =#        
        error("abc")        
    end
end

ExproniconLite.JLIfElse — Method

JLIfElse()

Create an emptry ifelse syntax type instance.

ExproniconLite.JLKwField — Type

JLKwField <: JLExpr

Type describes a Julia field that can have a default value in a Julia struct.

ExproniconLite.JLKwField — Method

JLKwField(;name, type=Any, doc=nothing, line=nothing, default=no_default)

Create a JLKwField instance.

ExproniconLite.JLKwStruct — Type

JLKwStruct <: JLExpr

Type describes a Julia struct that allows keyword definition of defaults.

ExproniconLite.JLKwStruct — Type

JLKwStruct(ex::Expr, typealias=nothing)

Create a JLKwStruct from given Julia struct Expr, with an option to attach an alias to this type name.

Example

julia> JLKwStruct(:(struct Foo
           x::Int = 1
       end))
#= kw =# struct Foo
    #= REPL[39]:2 =#
    x::Int = 1
end

ExproniconLite.JLKwStruct — Method

JLKwStruct(;kw...)

Create a JLKwStruct instance.

Kwargs

name: required struct name.

typealias: alias to the struct.

ismutable: if it is a mutable struct.

typevars: type variables.

supertype: supertype of the struct.

fields: fields of the struct, must be in type JLKwField.

constructors: constructors of the struct, must be in type JLFunction.

line: LineNumberNode of the struct.

doc: doc string of the struct.

misc: other expressions inside the struct expression.

ExproniconLite.JLStruct — Type

JLStruct <: JLExpr

Type describes a Julia struct.

Example

Construct a Julia struct.

julia> JLStruct(;name=:Foo, typevars=[:T], fields=[JLField(;name=:x, type=Int)])
struct Foo{T}
    x::Int64
end

Decompose a Julia struct expression

julia> ex = :(struct Foo{T}
           x::Int64
       end)
:(struct Foo{T}
      #= REPL[31]:2 =#
      x::Int64
  end)

julia> jl = JLStruct(ex)
struct Foo{T}
    #= REPL[31]:2 =#
    x::Int64
end

Generate a Julia struct expression

julia> codegen_ast(jl)
:(struct Foo{T}
      #= REPL[31]:2 =#
      x::Int64
  end)

ExproniconLite.JLStruct — Method

JLStruct(ex::Expr)

Create a JLStruct object from a Julia struct Expr.

Example

julia> JLStruct(:(struct Foo
           x::Int
       end))
struct Foo
    #= REPL[38]:2 =#
    x::Int
end

ExproniconLite.JLStruct — Method

JLStruct(;kw...)

Create a JLStruct instance.

Kwargs

name: required struct name.

ismutable: if it is a mutable struct.

typevars: type variables.

supertype: supertype of the struct.

fields: fields of the struct, must be in type JLField.

constructors: constructors of the struct, must be in type JLFunction.

line: LineNumberNode of the struct.

doc: doc string of the struct.

misc: other expressions inside the struct expression.

ExproniconLite.NoDefault — Type

NoDefault

Type describes a field should have no default value.

ExproniconLite.codegen_ast — Method

codegen_ast(def)

Generate Julia AST object Expr from a given syntax type.

Example

One can generate the Julia AST object from a JLKwStruct syntax type.

julia> def = @expr JLKwStruct struct Foo{N, T}
                  x::T = 1
              end
#= kw =# struct Foo{N, T}
    #= REPL[19]:2 =#
    x::T = 1
end

julia> codegen_ast(def)|>rm_lineinfo
quote
    struct Foo{N, T}
        x::T
    end
    begin
        function Foo{N, T}(; x = 1) where {N, T}
            Foo{N, T}(x)
        end
        function Foo{N}(; x::T = 1) where {N, T}
            Foo{N, T}(x)
        end
    end
end

ExproniconLite.codegen_ast_fields — Method

codegen_ast_fields(fields; just_name::Bool=true)

Generate a list of Julia AST object for each field, only generate a list of field names by default, option just_name can be turned off to call codegen_ast on each field object.

ExproniconLite.codegen_ast_kwfn — Function

codegen_ast_kwfn(def[, name = nothing])

Generate the keyword function from a Julia struct definition.

Example

julia> def = @expr JLKwStruct struct Foo{N, T}
                  x::T = 1
              end
#= kw =# struct Foo{N, T}
    #= REPL[19]:2 =#
    x::T = 1
end

julia> codegen_ast_kwfn(def)|>prettify
quote
    function Foo{N, T}(; x = 1) where {N, T}
        Foo{N, T}(x)
    end
    function Foo{N}(; x::T = 1) where {N, T}
        Foo{N, T}(x)
    end
end

julia> def = @expr JLKwStruct struct Foo
                  x::Int = 1
              end
#= kw =# struct Foo
    #= REPL[23]:2 =#
    x::Int = 1
end

julia> codegen_ast_kwfn(def)|>prettify
quote
    function Foo(; x = 1)
        Foo(x)
    end
    nothing
end

ExproniconLite.codegen_ast_kwfn_infer — Function

codegen_ast_kwfn_infer(def, name = nothing)

Generate the keyword function that infers the type.

ExproniconLite.codegen_ast_kwfn_plain — Function

codegen_ast_kwfn_plain(def[, name = nothing])

Generate the plain keyword function that does not infer type variables. So that one can use the type conversions defined by constructors.

ExproniconLite.codegen_ast_struct — Method

codegen_ast_struct(def)

Generate pure Julia struct Expr from struct definition. This is equivalent to codegen_ast for JLStruct. See also codegen_ast.

Example

julia> def = JLKwStruct(:(struct Foo
           x::Int=1
           
           Foo(x::Int) = new(x)
       end))
struct Foo
    x::Int = 1
end

julia> codegen_ast_struct(def)
:(struct Foo
      #= REPL[21]:2 =#
      x::Int
      Foo(x::Int) = begin
              #= REPL[21]:4 =#
              new(x)
          end
  end)

ExproniconLite.codegen_ast_struct_body — Method

codegen_ast_struct_body(def)

Generate the struct body.

Example

julia> def = JLStruct(:(struct Foo
           x::Int
           
           Foo(x::Int) = new(x)
       end))
struct Foo
    x::Int
end

julia> codegen_ast_struct_body(def)
quote
    #= REPL[15]:2 =#
    x::Int
    Foo(x::Int) = begin
            #= REPL[15]:4 =#
            new(x)
        end
end

ExproniconLite.codegen_ast_struct_head — Method

codegen_ast_struct_head(def)

Generate the struct head.

Example

julia> using Expronicon

julia> def = JLStruct(:(struct Foo{T} end))
struct Foo{T}
end

julia> codegen_ast_struct_head(def)
:(Foo{T})

julia> def = JLStruct(:(struct Foo{T} <: AbstractArray end))
struct Foo{T} <: AbstractArray
end

julia> codegen_ast_struct_head(def)
:(Foo{T} <: AbstractArray)

ExproniconLite.compare_expr — Method

compare_expr(lhs, rhs)

Compare two expression of type Expr or Symbol semantically, which:

ignore the detail value LineNumberNode in comparision

ignore the detailed name of typevars in Expr(:curly, ...) or Expr(:where, ...)

This gives a way to compare two Julia expression semantically which means although some details of the expression is different but they should produce the same lowered code.

ExproniconLite.compare_vars — Method

compare_vars(lhs, rhs)

Compare two expression by assuming all Symbols are variables, thus their value doesn't matter, only where they are matters under this assumption. See also compare_expr.

ExproniconLite.eval_interp — Method

eval_interp(m::Module, ex)

evaluate the interpolation operator in ex inside given module m.

ExproniconLite.eval_literal — Method

eval_literal(m::Module, ex)

Evaluate the literal values and insert them back to the expression. The literal value can be checked via is_literal.

ExproniconLite.flatten_blocks — Method

flatten_blocks(ex)

Remove hierachical expression blocks.

ExproniconLite.has_kwfn_constructor — Function

has_kwfn_constructor(def[, name = struct_name_plain(def)])

Check if the struct definition contains keyword function constructor of name. The constructor name to check by default is the plain constructor which does not infer any type variables and requires user to input all type variables. See also struct_name_plain.

ExproniconLite.has_plain_constructor — Function

has_plain_constructor(def, name = struct_name_plain(def))

Check if the struct definition contains the plain constructor of name. By default the name is the inferable name struct_name_plain.

Example

def = @expr JLKwStruct struct Foo{T, N}
    x::Int
    y::N

    Foo{T, N}(x, y) where {T, N} = new{T, N}(x, y)
end

has_plain_constructor(def) # true

def = @expr JLKwStruct struct Foo{T, N}
    x::T
    y::N

    Foo(x, y) = new{typeof(x), typeof(y)}(x, y)
end

has_plain_constructor(def) # false

the arguments must have no type annotations.

def = @expr JLKwStruct struct Foo{T, N}
    x::T
    y::N

    Foo{T, N}(x::T, y::N) where {T, N} = new{T, N}(x, y)
end

has_plain_constructor(def) # false

ExproniconLite.is_field — Method

is_field(ex)

Check if ex is a valid field expression.

ExproniconLite.is_field_default — Method

is_field_default(ex)

Check if ex is a <field expr> = <default expr> expression.

ExproniconLite.is_for — Method

is_for(ex)

Check if ex is a for loop expression.

ExproniconLite.is_function — Method

is_function(def)

Check if given object is a function expression.

ExproniconLite.is_ifelse — Method

is_ifelse(ex)

Check if ex is an if ... elseif ... else ... end expression.

ExproniconLite.is_kw_function — Method

is_kw_function(def)

Check if a given function definition supports keyword arguments.

ExproniconLite.is_literal — Method

is_literal(x)

Check if x is a literal value.

ExproniconLite.is_struct — Method

is_struct(ex)

Check if ex is a struct expression.

ExproniconLite.is_struct_not_kw_struct — Method

is_struct_not_kw_struct(ex)

Check if ex is a struct expression excluding keyword struct syntax.

ExproniconLite.name_only — Method

name_only(ex)

Remove everything else leaving just names, currently supports function calls, type with type variables, subtype operator <: and type annotation ::.

Example

julia> using Expronicon

julia> name_only(:(sin(2)))
:sin

julia> name_only(:(Foo{Int}))
:Foo

julia> name_only(:(Foo{Int} <: Real))
:Foo

julia> name_only(:(x::Int))
:x

ExproniconLite.prettify — Method

prettify(ex)

Prettify given expression, remove all LineNumberNode and extra code blocks.

Tips

the LineNumberNode inside macro calls won't be removed since the macrocall expression requires a LineNumberNode. See also issues/#9.

ExproniconLite.rm_annotations — Method

rm_annotations(x)

Remove type annotation of given expression.

ExproniconLite.rm_lineinfo — Method

rm_lineinfo(ex)

Remove LineNumberNode in a given expression.

Tips

the LineNumberNode inside macro calls won't be removed since the macrocall expression requires a LineNumberNode. See also issues/#9.

ExproniconLite.rm_nothing — Method

rm_nothing(ex)

Remove the constant value nothing in given expression ex.

ExproniconLite.split_doc — Method

split_doc(ex::Expr) -> line, doc, expr

Split doc string from given expression.

ExproniconLite.split_function — Method

split_function(ex::Expr) -> head, call, body

Split function head declaration with function body.

ExproniconLite.split_function_head — Method

split_function_head(ex::Expr) -> name, args, kw, whereparams

Split function head to name, arguments, keyword arguments and where parameters.

ExproniconLite.split_struct — Method

split_struct(ex::Expr) -> ismutable, name, typevars, supertype, body

Split struct definition head and body.

ExproniconLite.split_struct_name — Method

split_struct_name(ex::Expr) -> name, typevars, supertype

Split the name, type parameters and supertype definition from struct declaration head.

ExproniconLite.struct_name_plain — Method

struct_name_plain(def)

Plain constructor name. See also struct_name_without_inferable.

Example

julia> def = @expr JLKwStruct struct Foo{N, Inferable}
    x::Inferable = 1
end

julia> struct_name_plain(def)
:(Foo{N, Inferable})

ExproniconLite.struct_name_without_inferable — Method

struct_name_without_inferable(def; leading_inferable::Bool=true)

Constructor name that assume some of the type variables is inferred. See also struct_name_plain. The kwarg leading_inferable can be used to configure whether to preserve the leading inferable type variables, the default is true to be consistent with the default julia constructors.

Example

julia> def = @expr JLKwStruct struct Foo{N, Inferable}
    x::Inferable = 1
end

julia> struct_name_without_inferable(def)
:(Foo{N})

julia> def = @expr JLKwStruct struct Foo{Inferable, NotInferable}
    x::Inferable
end

julia> struct_name_without_inferable(def; leading_inferable=true)
:(Foo{Inferable, NotInferable})

julia> struct_name_without_inferable(def; leading_inferable=false)
:(Foo{NotInferable})

ExproniconLite.subtitute — Method

subtitute(ex::Expr, old=>new)

Subtitute the old symbol old with new.

ExproniconLite.support_default — Method

support_default(f)

Check if field type f supports default value.

ExproniconLite.xcall — Method

xcall(name, args...; kw...)

Create a function call to name.

ExproniconLite.xcall — Method

xcall(m::Module, name::Symbol, args...; kw...)

Create a function call to GlobalRef(m, name).

Tip

due to Revise/#616, to make your macro work with Revise, use the dot expression Expr(:., <module>, QuoteNode(<name>)) instead of GlobalRef.

ExproniconLite.xfirst — Method

xfirst(collection)

Create a function call expression to Base.first.

ExproniconLite.xiterate — Method

xiterate(it, st)

Create a function call expression to Base.iterate.

ExproniconLite.xiterate — Method

xiterate(it)

Create a function call expression to Base.iterate.

ExproniconLite.xlast — Method

xlast(collection)

Create a function call expression to Base.last.

ExproniconLite.xmap — Method

xmap(f, xs...)

Create a function call expression to Base.map.

ExproniconLite.xmapreduce — Method

xmapreduce(f, op, xs...)

Create a function call expression to Base.mapreduce.

ExproniconLite.xnamedtuple — Method

xnamedtuple(;kw...)

Create a NamedTuple expression.

ExproniconLite.xprint — Method

xprint(xs...)

Create a function call expression to Base.print.

ExproniconLite.xprintln — Method

xprintln(xs...)

Create a function call expression to Base.println.

ExproniconLite.xpush — Method

xpush(collection, items...)

Create a function call expression to Base.push!.

ExproniconLite.xtuple — Method

xtuple(xs...)

Create a Tuple expression.

ExproniconLite.@expr — Macro

@expr <type> <expression>

Return the expression in given type.

Example

julia> ex = @expr JLKwStruct struct Foo{N, T}
           x::T = 1
       end
#= kw =# struct Foo{N, T}
    #= /home/roger/code/julia/Expronicon/test/analysis.jl:5 =#
    x::T = 1
end

ExproniconLite.@expr — Macro

@expr <expression>

Return the original expression object.

Example

julia> ex = @expr x + 1
:(x + 1)

ExproniconLite.@test_expr — Macro

@test_expr <type> <ex>

Test if the syntax type generates the same expression ex. Returns the corresponding syntax type instance. Requires using Test before using this macro.

Example

def = @test_expr JLFunction function (x, y)
    return 2
end
@test is_kw_fn(def) == false

ExproniconLite.@test_expr — Macro

@test_expr <expr> == <expr>

Test if two expression is equivalent semantically, this uses compare_expr to decide if they are equivalent, ignores things such as LineNumberNode generated Symbol in Expr(:curly, ...) or Expr(:where, ...).